+++ tests/framebuffer_test.cpp
@@ -23,12 +23,12 @@ TEST_CASE( "Renderers - Framebuffer test", "[manual]" )
     CTextRenderer textRenderer;
     REQUIRE_NOTHROW( textRenderer.init() );
     textRenderer.setViewMatrix( mat4f() );
-    textRenderer.setProjectionMatrix( matMakeOrthographic<float>( 0, window.getSizeWidth(), window.getSizeHeight(), 0, -1, 1 ) ); 
+    textRenderer.setProjectionMatrix( matMakeOrthographic<float>( 0, (float)window.getSizeWidth(), (float)window.getSizeHeight(), 0, -1, 1 ) ); 
 
     CSpriteRenderer spriteRenderer;
     REQUIRE_NOTHROW( spriteRenderer.init() );
     spriteRenderer.setViewMatrix( mat4f() );
-    spriteRenderer.setProjectionMatrix( matMakeOrthographic<float>( 0, window.getSizeWidth(), window.getSizeHeight(), 0, -1, 1 ) ); 
+    spriteRenderer.setProjectionMatrix( matMakeOrthographic<float>( 0, (float)window.getSizeWidth(), (float)window.getSizeHeight(), 0, -1, 1 ) ); 
 
 
     showInfoMessageBox( testName, "This test should show the same results received through direct and indirect (through target texture) rendering of text." );
